+++ libs/design/toast/src/service/changes-focus.ts
@@ -1,3 +1,3 @@
-import { DaffToast } from '../toast';
+import { DaffToast } from '../interfaces/toast';
 
 export const daffToastChangesFocus = (toast: DaffToast): boolean => toast.actions?.length > 0;
